Finish uniprot parsing
- finish parsing of uniprot data
- add option for skipping DB downloads

+++ database/helpers/download-ftp.js
@@ -1,25 +1,29 @@
 const Ftp = require('ftp');
 const fs = require('fs');
 
-const downloadFile = (host, file, dest) => (
+const downloadFile = (host, file, dest, skip) => (
   new Promise((resolve, reject) => {
-    const outFile = fs.createWriteStream(dest);
-    const client = new Ftp();
-    client.on('ready', () => {
-      client.get(file, (err, stream) => {
-        if (err) {
-          fs.unlink(dest);
-          reject(err);
-        }
-        stream.once('close', () => {
-          client.end();
-          outFile.close();
-          resolve();
+    if (skip) {
+      resolve();
+    } else {
+      const outFile = fs.createWriteStream(dest);
+      const client = new Ftp();
+      client.on('ready', () => {
+        client.get(file, (err, stream) => {
+          if (err) {
+            fs.unlink(dest);
+            reject(err);
+          }
+          stream.once('close', () => {
+            client.end();
+            outFile.close();
+            resolve();
+          });
+          stream.pipe(outFile);
         });
-        stream.pipe(outFile);
       });
-    });
-    client.connect({ host });
+      client.connect({ host });
+    }
   })
 );
